Particle starttime: uint32_t → float (fix Nightingale black desktop)
WPParticleObject::starttime was uint32_t.  Nightingale
(3276911872) ships deku_twinkle_shootingstar.json with
"starttime": -2, which wrapped to 4294967294.  MaxStartTime()
returned ~4.29e9 and PreSimulate — called synchronously at scene
load right after compileRenderGraph — tried to iterate ~134
billion ticks before the first draw ran.  Plasma's compositor
timed out and showed a black desktop.

Fix: float throughout the particle pipeline.  Negative values
stay negative; MaxStartTime's std::max(0, ...) accumulator clamps
them to 0, so PreSimulate early-returns.  Also preserves
sub-second values (0.69999999 no longer truncates to 0).

Two new ParticleSystem.StartTime test cases: negative-starttime
regression and fractional-starttime preservation.  1100 backend
tests pass.

+++ src/Particle/ParticleSystem.cpp
@@ -42,7 +42,7 @@ ParticleInstance::BoundedData& ParticleInstance::GetBoundedData() { return m_bou
 ParticleSubSystem::ParticleSubSystem(ParticleSystem& p, std::shared_ptr<SceneMesh> sm,
                                      uint32_t maxcount, double rate, u32 maxcount_instance,
                                      double probability, SpawnType type,
-                                     ParticleRawGenSpecOp specOp, uint32_t starttime)
+                                     ParticleRawGenSpecOp specOp, float starttime)
     : m_sys(p),
       m_mesh(sm),
       m_maxcount(maxcount),
